+++ src/components/common/modal/core/ModalProvider.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { useState } from 'react';
-import ModalContext from '@/components/common/modal/ModalContext';
+import ModalContext from '@/components/common/modal/core/ModalContext';
 
 export default function ModalProvider({ children }: { children: React.ReactNode }) {
   const [isOpen, setIsOpen] = useState(false);
